♻️ Refactor: Root → App으로 통합 및 Splash 처리 일원화
