+++ src/main.tsx
@@ -2,12 +2,13 @@ import './index.css';
 
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
+import { RouterProvider } from 'react-router-dom';
 
-import App from './App';
 import router from './routes/router';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App router={router} />
+    <RouterProvider router={router} />
+    {/* <App router={router} /> */}
   </StrictMode>,
 );
